# Fast path in weak head reduction of applied atoms.