@SkySkimmer 2017-04-08
Why not the whd_foo in checker/reduction.ml too?