@ppedrot 2017-04-08
Updated the patch according to your comments.