@silene 2017-04-09
There are a few additional cheap tests you could consider: fixpoints without enough arguments, match on a irreducible value, and so on. My main concern is that all these tests cause code bloat yet these situations might not occur in practice. But on the other hand, it feels like not fixing them right now is just waiting for a user to stumble on them, now that we know it is a source of inefficiency. So I am not sure what the way forward is.