@ppedrot 2017-04-09
Timings:
```

┌────────────────────────┬─────────────────────────┬─────────────────────────────────────┬───────────────────────────────────────┐
│                        │      user time [s]      │             CPU cycles              │           CPU instructions            │
│                        │                         │                                     │                                       │
│           package_name │     NEW     OLD PDIFF   │           NEW           OLD PDIFF   │            NEW            OLD PDIFF   │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│              coq-flocq │   54.98   55.23 -0.45 % │  150107120229  150905091050 -0.53 % │   192821866714   194097218211 -0.66 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│             coq-geocoq │ 2229.69 2238.98 -0.41 % │ 6207066443054 6234922639542 -0.45 % │ 10415557975765 10417799401878 -0.02 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│               coq-hott │  230.76  231.72 -0.41 % │  622052983211  624915418613 -0.46 % │  1047128705346  1048493305973 -0.13 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│               coq-iris │  431.44  432.99 -0.36 % │ 1196459895781 1199623991128 -0.26 % │  1729626336780  1729763797770 -0.01 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│              coq-color │  640.89  643.00 -0.33 % │ 1777577477248 1780754553296 -0.18 % │  2291380683022  2291762793669 -0.02 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│            coq-unimath │ 1223.08 1226.65 -0.29 % │ 3391203338022 3399659585341 -0.25 % │  5731538325536  5729033344855 +0.04 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│        coq-fiat-crypto │ 2497.39 2504.10 -0.27 % │ 6937741445945 6954434670873 -0.24 % │ 11762447714167 11794988900380 -0.28 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│                coq-vst │ 1490.69 1493.81 -0.21 % │ 4138876967576 4148402544226 -0.23 % │  6252803790732  6256063040826 -0.05 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│           coq-compcert │  806.29  807.45 -0.14 % │ 2240494609972 2239726244341 +0.03 % │  3407507158714  3408480713697 -0.03 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│   coq-mathcomp-algebra │  164.30  164.45 -0.09 % │  454685938555  454922643994 -0.05 % │   651518706364   651574891164 -0.01 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│     coq-mathcomp-field │  444.24  444.32 -0.02 % │ 1236145995142 1236577636574 -0.03 % │  2083435770616  2083578709459 -0.01 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│       coq-fiat-parsers │  527.52  527.58 -0.01 % │ 1442804219342 1445006587189 -0.15 % │  2051556060253  2052601208886 -0.05 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│ coq-mathcomp-odd_order │ 1627.70 1626.46 +0.08 % │ 4534755197225 4524163063457 +0.23 % │  8130371630569  8129084144638 +0.02 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│         coq-coquelicot │   71.07   70.93 +0.20 % │  194714125933  194812744146 -0.05 % │   245553162452   245882711762 -0.13 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│ coq-mathcomp-character │  257.13  256.61 +0.20 % │  713730118616  713829002874 -0.01 % │  1060183107305  1060081576116 +0.01 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│  coq-mathcomp-solvable │  192.62  192.23 +0.20 % │  533298089073  533619032994 -0.06 % │   793795798501   794000711453 -0.03 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│ coq-mathcomp-ssreflect │   41.85   41.75 +0.24 % │  113801300926  113824286590 -0.02 % │   147136630873   147317953464 -0.12 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│  coq-mathcomp-fingroup │   55.43   55.26 +0.31 % │  152099658843  152129494221 -0.02 % │   214634160148   214985413727 -0.16 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│                 coq-sf │   43.79   43.62 +0.39 % │  119349456355  119352672693 -0.00 % │   156141196098   156234222480 -0.06 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│    coq-mathcomp-finmap │   34.92   34.70 +0.63 % │   96563725534   96413407635 +0.16 % │   147939036142   147908911783 +0.02 % │
└────────────────────────┴─────────────────────────┴─────────────────────────────────────┴───────────────────────────────────────┘
NEW = aa704399c4d4b8a74f4d6f42e65808c1ceab3b7e
OLD = 2c0287fe8445bd4b599bf8498bcb71b2a7df0d51
```