@ppedrot 2017-04-09
@silene I don't think those are critical. The real source of inefficiency in the bug I mentioned is the fact people use `whd_all` to get a restricted subset of weak-head normal terms, typically (applied) constructors or inductive types, and by typing assumption, it is known that they will evaluate to something like this. 

As suggested by @maximedenes, we should review the code and categorize uses of reduction functions, but I doubt people are actually expecting to recover more general neutral terms like blocked matches and fixpoints.