@maximedenes 2017-04-10
This looks like a good first step: the code is still readable, and the cases it covers may happen in practice. @silene, are you ok with this being merged as it is?