@silene 2017-04-11
> are you ok with this being merged as it is?

Yes.